feat(cor-handling): implement CORS handling utilities and enhance image download functionality with retry logic
